Revert "move stuff to qt"
This reverts commit aa2cd61633017dfcf35bdf97bda999f816aead8c.
