+++ napari/utils/notebook_display.py
@@ -1,6 +1,6 @@
 from io import BytesIO
 
-__all__ = ['NotebookScreenshot']
+__all__ = ['nbscreenshot']
 
 
 class NotebookScreenshot:
@@ -67,3 +67,6 @@ class NotebookScreenshot:
             file_obj.seek(0)
             png = file_obj.read()
         return png
+
+
+nbscreenshot = NotebookScreenshot
